Move some stuff around between files, add new dev dependency, updated scipy dependency
