+++ README.md
@@ -1,20 +1,56 @@
+## Continuous Integration Status
 [![DeepSource](https://deepsource.io/gh/Bluesy1/CharB0T.svg/?label=active+issues&show_trend=true&token=wb08Mu693aFli6scALkkEmql)](https://deepsource.io/gh/Bluesy1/CharB0T/?ref=repository-badge)
 [![DeepSource](https://deepsource.io/gh/Bluesy1/CharB0T.svg/?label=resolved+issues&show_trend=true&token=wb08Mu693aFli6scALkkEmql)](https://deepsource.io/gh/Bluesy1/CharB0T/?ref=repository-badge)
-[![CodeQL](https://github.com/Bluesy1/CharB0T/actions/workflows/codeql-analysis.yml/badge.svg)](https://github.com/Bluesy1/CharB0T/actions/workflows/codeql-analysis.yml)
-[![Code style: black](https://img.shields.io/badge/code%20style-black-000000.svg)](https://github.com/psf/black)
-[![Checked with pyright](https://img.shields.io/badge/pyright-checked-informational.svg)](https://github.com/microsoft/pyright/)
-[![Lint](https://github.com/Bluesy1/CharB0T/actions/workflows/lint.yml/badge.svg)](https://github.com/Bluesy1/CharB0T/actions/workflows/lint.yml)
 [![CodeFactor](https://www.codefactor.io/repository/github/bluesy1/charb0t/badge)](https://www.codefactor.io/repository/github/bluesy1/charb0t)
-[![](https://tokei.rs/b1/github/Bluesy1/CharB0T?category=code)](https://github.com/XAMPPRocky/tokei)
-[![](https://tokei.rs/b1/github/Bluesy1/CharB0T?category=lines)](https://github.com/XAMPPRocky/tokei)
-[![](https://tokei.rs/b1/github/Bluesy1/CharB0T?category=files)](https://github.com/XAMPPRocky/tokei)
-[![](https://tokei.rs/b1/github/Bluesy1/CharB0T?category=comments)](https://github.com/XAMPPRocky/tokei)
-![](https://img.shields.io/badge/license-MIT-blue.svg)
-[![wakatime](https://wakatime.com/badge/github/Bluesy1/CharB0T.svg)](https://wakatime.com/badge/github/Bluesy1/CharB0T)
-# CharB0T
-Personal Bots, And Ones made for Charlie.
 
-You are welcome to take any and adapt them, but it is not a usecase I support, so it is unlikely I will provide support for them. I do support discussions about how the code works and I can improve it or make it more efficient.
+[![Lint](https://github.com/Bluesy1/CharB0T/actions/workflows/lint.yml/badge.svg)](https://github.com/Bluesy1/CharB0T/actions/workflows/lint.yml)
+[![CodeQL](https://github.com/Bluesy1/CharB0T/actions/workflows/codeql-analysis.yml/badge.svg)](https://github.com/Bluesy1/CharB0T/actions/workflows/codeql-analysis.yml)
+[![pages-build-deployment](https://github.com/Bluesy1/CharB0T/actions/workflows/pages/pages-build-deployment/badge.svg?branch=gh-pages)](https://github.com/Bluesy1/CharB0T/actions/workflows/pages/pages-build-deployment)
+
+## Code Info
+
+![](https://img.shields.io/badge/license-MIT-blue.svg)
+[![Checked with pyright](https://img.shields.io/badge/pyright-checked-informational.svg)](https://github.com/microsoft/pyright/)
+[![Code style: black](https://img.shields.io/badge/code%20style-black-000000.svg)](https://github.com/psf/black)
+[![Code style: flake8](https://img.shields.io/badge/code%20style-flake8-blue.svg)](https://github.com/pycqa/flake8)
+[![wakatime](https://wakatime.com/badge/github/Bluesy1/CharB0T.svg)](https://wakatime.com/badge/github/Bluesy1/CharB0T)
+
+![GitHub commit activity](https://img.shields.io/github/commit-activity/w/Bluesy1/CharB0T)
+![GitHub issues by-label](https://img.shields.io/github/issues/Bluesy1/CharB0T/feature%20request?label=Feature%20Requests)
+![GitHub issues by-label](https://img.shields.io/github/issues/Bluesy1/CharB0T/bug?label=Confirmed%20Bugs)
+![GitHub issues by-label](https://img.shields.io/github/issues/Bluesy1/CharB0T/unconfirmed%20bug?label=Unconfirmed%20Bugs)
+![GitHub issues](https://img.shields.io/github/issues/Bluesy1/CharB0T)
+![GitHub closed issues](https://img.shields.io/github/issues-closed/Bluesy1/CharB0T)
+![GitHub pull requests](https://img.shields.io/github/issues-pr/Bluesy1/CharB0T)
+![GitHub closed pull requests](https://img.shields.io/github/issues-pr-closed/Bluesy1/CharB0T)
+
+## Contributors
+
+![GitHub contributors](https://img.shields.io/github/contributors/Bluesy1/CharB0T.svg)
+[![Discord](https://img.shields.io/discord/225345178955808768)](https://cpry.net/discord)
+
+![Contributors](https://contrib.rocks/image?repo=BLuesy1/CharB0T)
+
+## Docs
+
+You can find the documentation for the bot [here](https://bluesy1.github.io/CharB0T/docs/).
+The changelog is [here](https://bluesy1.github.io/CharB0T/changes).
+
+## Contributing
+
+- If you have a good idea for a new feature, create a [Feature Request](https://github.com/Bluesy1/CharB0T/issues/new?assignees=&labels=feature+request&template=feature_request.yml).
+- If you encounter a bug, create a [Bug Report](https://github.com/Bluesy1/CharB0T/issues/new?assignees=&labels=unconfirmed+bug&template=bug_report.yml).
+
+
+1. Check for an open issue that hasn't been resolved yet.
+2. If there is an open issue, check if it is a bug or a feature request.
+3. If it is a bug that has been confirmed, if it hasn't been assigned, assign it to yourself, fork the repository,
+  fix the bug, and submit a pull request.
+4. If it is an unconfirmed bug, see if you can reproduce it consistently, and report that. If you can't reproduce it,
+  also report it. If you can reproduce it, see if you can fix it, and create a PR through a fork.
+5. If it is a feature request, if it has been approved by Bluesy and you can implement it, create a PR through a fork.
+6. If you implement or change something documented, update the documentation and changelogs and submit a pull request for that too.
+7. Again, **Be sure to update the changelogs and documentation.** If you don't, I will have to write them myself and your PR will be delayed. -
 
 ----------------------------------------------------------------------------
 MIT License
