+++ charbot/giveaway.py
@@ -536,77 +536,6 @@ class Giveaway(commands.Cog):
         """
         return pd.read_csv("giveaway.csv", index_col=0, usecols=[0, 1, 2, 4], names=["date", "game", "url", "source"])
 
-    @app_commands.command(name="rollcall", description="Claim your daily reputation bonus")
-    @app_commands.guilds(225345178955808768)
-    async def rollcall(self, interaction: discord.Interaction):
-        """Get a daily reputation bonus.
-
-        Parameters
-        ----------
-        interaction: discord.Interaction
-            The interaction of the command invocation.
-        """
-        user = interaction.user
-        clientuser = self.bot.user
-        assert isinstance(clientuser, discord.ClientUser)  # skipcq: BAN-B101
-        assert isinstance(user, discord.Member)  # skipcq: BAN-B101
-        if not any(role.id in ALLOWED_ROLES for role in user.roles) or interaction.channel_id != CHANNEL_ID:
-            await interaction.response.send_message(MESSAGE, ephemeral=True)
-            return
-        await interaction.response.defer(ephemeral=True)
-        giveaway_user = await self.bot.giveaway_user(user.id)
-        if giveaway_user is None:
-            async with self.bot.pool.acquire() as conn:
-                await conn.execute("INSERT INTO users (id, points) VALUES ($1, 20)", user.id)
-                await conn.execute(
-                    "INSERT INTO daily_points (id, last_claim, last_particip_dt, particip, won) VALUES "
-                    "($1, $2, $3, 0, 0)",
-                    user.id,
-                    self.bot.TIME(),
-                    self.bot.TIME() - datetime.timedelta(days=1),
-                )
-                await conn.execute("INSERT INTO bids (id, bid) VALUES ($1, 0)", user.id)
-                await interaction.followup.send("You got some Rep today, inmate")
-                await self.bot.program_logs.send(
-                    f"{user.mention} has claimed their daily reputation bonus.",
-                    allowed_mentions=discord.AllowedMentions(users=False),
-                    username=clientuser.name,
-                    avatar_url=clientuser.display_avatar.url,
-                )
-            return
-        if giveaway_user["daily"] >= self.bot.TIME():
-            await interaction.followup.send("No more Rep for you yet, get back to your cell")
-            return
-        async with self.bot.pool.acquire() as conn:
-            await conn.execute("UPDATE users SET points = points + 20 WHERE id = $1", user.id)
-            await conn.execute("UPDATE daily_points SET last_claim = $1 WHERE id = $2", self.bot.TIME(), user.id)
-        await interaction.followup.send("You got some Rep today, inmate")
-        await self.bot.program_logs.send(
-            f"{user.mention} has claimed their daily reputation bonus.",
-            allowed_mentions=discord.AllowedMentions(users=False),
-            username=clientuser.name,
-            avatar_url=clientuser.display_avatar.url,
-        )
-
-    @app_commands.command(name="reputation", description="Check your reputation")
-    @app_commands.guilds(225345178955808768)
-    async def query_points(self, interaction: discord.Interaction):
-        """Query your reputation.
-
-        Parameters
-        ----------
-        interaction: discord.Interaction
-            The interaction of the command invocation.
-        """
-        user = interaction.user
-        assert isinstance(user, discord.Member)  # skipcq: BAN-B101
-        if not any(role.id in ALLOWED_ROLES for role in user.roles) or interaction.channel_id != CHANNEL_ID:
-            await interaction.response.send_message(MESSAGE, ephemeral=True)
-            return
-        await interaction.response.defer(ephemeral=True)
-        points = await self.bot.pool.fetchval("SELECT points from users where id = $1", user.id) or 0
-        await interaction.followup.send(f"You have {points} reputation.", ephemeral=True)
-
     @app_commands.command(name="confirm", description="[Charlie only] confirm a winner")
     @app_commands.describe(user="The user to confirm as a winner.")
     @app_commands.guilds(225345178955808768)
