+++ dev-requirements.txt
@@ -3,6 +3,7 @@ bandit==1.7.4
 black==22.3.0
 flake8==4.0.1
 isort==5.10.1
+pre-commit==2.19.0
 pydocstyle==6.1.1
 pyright==1.1.247
 toml==0.10.2
