+++ requirements.txt
@@ -1,4 +1,4 @@
-aiohttp==3.8.1
+aiohttp[speedups]==3.8.1
 asyncpg==0.25.0
 git+https://github.com/Rapptz/discord.py@master#egg=discord.py
 jishaku==2.5.0
@@ -9,7 +9,7 @@ pandas==1.4.2
 Pillow==9.1.1
 python-dotenv==0.20.0
 pytz==2022.1
-scipy==1.8.0
+scipy==1.8.1
 uvloop==0.16.0
 validators==0.19.0
 yfinance==0.1.70
